# api.yaml
openapi: 3.0.3
info:
  title: Order Service API
  description: REST API for order management. Use JWT-token from Authentication Service.
  version: 1.0.0
servers:
  - url: http://localhost:8082
    description: Local development server
tags:
  - name: Orders
    description: Order CRUD and query operations
paths:
  /api/orders/{id}:
    get:
      tags:
        - Orders
      operationId: getOrderById
      summary: Get order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Order found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderWithUserResponse'
        '404':
          description: Order not found
    put:
      tags:
        - Orders
      operationId: updateOrder
      summary: Update order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderDto'
      responses:
        '200':
          description: Order updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderWithUserResponse'
        '404':
          description: Order not found
    delete:
      tags:
        - Orders
      operationId: deleteOrder
      summary: Delete order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Order deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDto'
        '404':
          description: Order not found
  /api/orders:
    post:
      tags:
        - Orders
      operationId: createOrder
      summary: Create a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderDto'
      responses:
        '200':
          description: Order created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderWithUserResponse'
        '404':
          description: Order could not be created
  /api/orders/by-email:
    get:
      tags:
        - Orders
      operationId: getOrdersByUserEmail
      summary: Get orders by user email
      parameters:
        - name: email
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderWithUserResponse'
  /api/orders/find-by-ids:
    get:
      tags:
        - Orders
      operationId: getOrdersIdIn
      summary: Get orders by IDs
      parameters:
        - name: ids
          in: query
          required: true
          schema:
            type: array
            items:
              type: string
              format: uuid
          style: form
          explode: true
      responses:
        '200':
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderWithUserResponse'
  /api/orders/find-by-statuses:
    get:
      tags:
        - Orders
      operationId: getByStatusIn
      summary: Get orders by statuses
      parameters:
        - name: statuses
          in: query
          required: true
          schema:
            type: array
            items:
              $ref: '#/components/schemas/OrderStatus'
          style: form
          explode: true
      responses:
        '200':
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderWithUserResponse'
  /api/orders/all:
    get:
      tags:
        - Orders
      operationId: getAllOrders
      summary: Get all orders
      responses:
        '200':
          description: List of all orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderWithUserResponse'
  /api/orders/paginated:
    get:
      tags:
        - Orders
      operationId: getAllOrdersWithPagination
      summary: Get paginated orders
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Paginated orders
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageOrderDto'
components:
  schemas:
    OrderStatus:
      type: string
      enum:
        - CREATED
        - PROCESSING
        - PAID
        - CANCELLED
        - FAILED
    OrderItemDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        orderId:
          type: string
          format: uuid
        itemId:
          type: integer
          format: int64
        quantity:
          type: integer
          format: int64
    OrderDto:
      type: object
      required:
        - userId
        - orderItems
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: integer
          format: int64
        status:
          $ref: '#/components/schemas/OrderStatus'
        creationDate:
          type: string
          format: date-time
        orderItems:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemDto'
    UserDto:
      type: object
      properties:
        userId:
          type: integer
          format: int64
        name:
          type: string
        surname:
          type: string
        birthDate:
          type: string
          format: date
        email:
          type: string
    OrderWithUserResponse:
      type: object
      properties:
        order:
          $ref: '#/components/schemas/OrderDto'
        user:
          $ref: '#/components/schemas/UserDto'
    PageOrderDto:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/OrderDto'
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        size:
          type: integer
        number:
          type: integer
        first:
          type: boolean
        last:
          type: boolean
        empty:
          type: boolean
        numberOfElements:
          type: integer